Add epochs and subsumes to yaml schema
Note: I had added epochs to the copy of this file in https://github.com/familysearch/gedcom-registries when we first added it, but forgot to update this it in this repository. We should probably reduce the number of copies of yaml-schema.yaml that we need to maintain to just one.