YAML format should allow tracking applications using extensions
https://github.com/familysearch/GEDCOM-registries?tab=readme-ov-file#registering-extensions
says that "once an extension meets the Valuable, Absent, and Used
criteria then a standard tag/URI can be added...".

https://github.com/FamilySearch/GEDCOM/tree/main/attribute-event-requests
tracks what things are used by apps, but it's only for attributes and
events, not other extensions.

This PR proposes the ability to tracking "used by" in any YAML file
that describes an extension.

Signed-off-by: Dave Thaler <dthaler@armidalesoftware.com>