Add HEAD.NOTE to many example files (#206)
As noted in https://github.com/FamilySearch/GEDCOM/issues/561 many of the example files don't make any sense:
they contain more-or-less random payloads of the right type for each structure. This adds the following to the header of each such file:

```ged
1 NOTE This file is intended to provide coverage of parts of the specification and does not contain meaningful historical or genealogical data.
```

Note that some example files are made-up data but at lease somewhat internally consistent; for example, `same-sex-marriage.ged` has internally-consistent (but made-up) names and relationships. This does not add that note to those files.

This commit also does not change the .gdz files. I can't recall if those are automatically updated or if I need to update them manually.