FAQ on data security (#278)
* FAQ on data security

Resolve #256 

It was far from clear to me how much detail to provide here. I hope this is the right mix of informative but not overboard.

Note that this entry does not contain a recommendation, just a discussion. I toyed with adding some kind of explicit acknowledgement of that fact at the end, some kind of "we can't make a recommendation, you have to think about this yourself" but wasn't sure what to say so didn't add one.

* typos

* Update _pages/techfaqs.md

* Update _pages/techfaqs.md

Co-authored-by: Dave Thaler <dthaler1968@gmail.com>

---------

Co-authored-by: Dave Thaler <dthaler1968@gmail.com>